requirements/base.txt
django
psycopg2
django-braces
django-crispy-forms
django-solo
pillow
reportlab